Chat Agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.14
langchain-google-genai==0.0.5
google-generativeai==0.3.2
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2
llama_cpp